Convert akka-samples-persistence-scale to pekko:
Update dirname

Update build.sbt file

Update application config. Note TODO: update URL to point at MGMT docs. Update pekko mgmt port?

Update source code to use pekko

Update docs
