--- akka-sample-persistence-dc-java/pom.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-   <modelVersion>4.0.0</modelVersion>
-  <artifactId>akka-sample-replicated-event-sourcing-java</artifactId>
-  <groupId>com.lightbend.akka.samples</groupId>
-  <name>Akka replicated event sourcing multi dc sample</name>
-  <version>1.0</version>
-
-   <licenses>
-    <license>
-      <name>Public Domain (CC0)</name>
-      <url>http://creativecommons.org/publicdomain/zero/1.0/</url>
-    </license>
-  </licenses>
-
-  <repositories>
-  <repository>
-    <id>lightbend-commercial</id>
-    <name>Lightbend Commercial</name>
-    <url>https://lightbend.bintray.com/commercial-releases</url>
-   </repository>
- </repositories>
-
-   <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <akka.version>2.6.20</akka.version>
-    <akka-persistence-cassandra.version>1.0.6</akka-persistence-cassandra.version>
-    <akka-http.version>10.2.10</akka-http.version>
-    <akka-management.version>1.1.4</akka-management.version>
-  </properties>
-
-   <dependencies>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-cluster-sharding-typed_2.13</artifactId>
-      <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-        <groupId>com.typesafe.akka</groupId>
-        <artifactId>akka-cluster_2.13</artifactId>
-        <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-        <groupId>com.typesafe.akka</groupId>
-        <artifactId>akka-actor_2.13</artifactId>
-        <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-        <groupId>com.typesafe.akka</groupId>
-        <artifactId>akka-cluster-tools_2.13</artifactId>
-        <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-        <groupId>com.typesafe.akka</groupId>
-        <artifactId>akka-serialization-jackson_2.13</artifactId>
-        <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-persistence-typed_2.13</artifactId>
-      <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-persistence-query_2.13</artifactId>
-      <version>${akka.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-persistence-cassandra_2.13</artifactId>
-      <version>${akka-persistence-cassandra.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-persistence-cassandra-launcher_2.13</artifactId>
-      <version>${akka-persistence-cassandra.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-http_2.13</artifactId>
-      <version>${akka-http.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-http-spray-json_2.13</artifactId>
-      <version>${akka-http.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.lightbend.akka.management</groupId>
-      <artifactId>akka-management_2.13</artifactId>
-      <version>${akka-management.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.lightbend.akka.management</groupId>
-      <artifactId>akka-management-cluster-http_2.13</artifactId>
-      <version>${akka-management.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>ch.qos.logback</groupId>
-      <artifactId>logback-classic</artifactId>
-      <version>1.2.11</version>
-    </dependency>
-     <dependency>
-      <groupId>com.typesafe.akka</groupId>
-      <artifactId>akka-actor-testkit-typed_2.13</artifactId>
-      <version>${akka.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.13.1</version>
-      <scope>test</scope>
-    </dependency>
-   </dependencies>
-
-   <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-          <compilerArgs>
-            <arg>-Xlint:unchecked</arg>
-            <arg>-Xlint:deprecation</arg>
-            <arg>-parameters</arg>
-          </compilerArgs>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
- </project>
