--- akka-sample-persistence-dc-java/src/main/java/sample/persistence/res/CborSerializable.java
@@ -1,4 +1,0 @@
-package sample.persistence.res;
-
-public interface CborSerializable {
-}
