--- akka-sample-persistence-dc-java/src/test/resources/application-test.conf
@@ -1,7 +1,0 @@
-akka {
-  actor {
-    serialization-bindings {
-      "sample.persistence.res.CborSerializable" = jackson-cbor
-    }
-  }
-}
