--- akka-sample-persistence-dc-scala/src/main/resources/logback.xml
@@ -1,24 +1,0 @@
-<configuration>
-    <appender name="STDOUT" target="System.out" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>[%date{ISO8601}] [%level] [%logger] [%marker] [%thread] - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <appender name="ASYNC" class="ch.qos.logback.classic.AsyncAppender">
-        <queueSize>8192</queueSize>
-        <neverBlock>true</neverBlock>
-        <appender-ref ref="STDOUT" />
-    </appender>
-
-    <logger name="com.datastax.driver.core.QueryLogger.NORMAL" level="INFO"/>
-    <logger name="com.datastax.driver.core.QueryLogger.ERROR" level="DEBUG"/>
-    <logger name="com.datastax.driver.core.QueryLogger.SLOW" level="DEBUG"/>
-    <logger name="com.datastax.driver.core.Connection" level="INFO"/>
-    <logger name="com.codahale.metrics" level="INFO"/>
-
-    <root level="INFO">
-        <appender-ref ref="ASYNC"/>
-    </root>
-
-</configuration>
