+++ pekko-sample-persistence-dc-scala/project/build.properties
