+++ pekko-sample-persistence-dc-scala/project/plugins.sbt
