+++ pekko-sample-persistence-dc-scala/src/main/resources/application.conf
@@ -1,0 +1,59 @@
+pekko {
+  loglevel = INFO
+
+  actor {
+    provider = cluster
+
+    serialization-bindings {
+      "sample.persistence.res.CborSerializable" = jackson-cbor
+    }
+  }
+
+    # For the sample, just bind to loopback and do not allow access from the network
+    # the port is overridden by the logic in main class
+    remote.artery {
+      canonical.port = 0
+      canonical.hostname = 127.0.0.1
+    }
+
+    cluster {
+      seed-nodes = [
+        "pekko://ClusterSystem@127.0.0.1:7345",
+        "pekko://ClusterSystem@127.0.0.1:7355"]
+
+      # Needed when running many actor systems in the same JVM
+      jmx.multi-mbeans-in-same-jvm = on
+
+      downing-provider-class = "pekko.cluster.sbr.SplitBrainResolverProvider"
+
+      multi-data-center {
+        self-data-center = "eu-west"
+      }
+    }
+
+    persistence {
+      journal.plugin = "pekko.persistence.cassandra.journal"
+    }
+
+  }
+
+# don't do this in production, convenient for local example
+pekko.persistence.cassandra {
+ journal {
+   keyspace-autocreate = true
+   tables-autocreate = true
+ }
+ snapshot {
+   keyspace-autocreate = true
+   tables-autocreate = true
+ }
+}
+
+datastax-java-driver.advanced.reconnect-on-init = true
+
+
+# Apache Pekko Management config: https://developer.lightbend.com/docs/akka-management/current/index.html
+pekko.management {
+  http.hostname = "127.0.0.1"
+  http.port = 19999
+}
