+++ pekko-sample-persistence-dc-scala/src/main/scala/sample/persistence/res/CborSerializable.scala
